// src/app/core/enums/route-name.enum.ts
export enum RouteName {
    Dashboard = 'dashboard',
    Utilisateurs = 'utilisateurs',
    Agence = 'Agence',
    Clients = 'clients',
    Suppliers = 'fournisseurs',
    Produits = 'produits',
    Parameters = 'parameters',
    Dossier = 'dossier',
    CategoryDocument = 'categorie-document',
    Messaging = 'messagerie',
    Numeration = 'numerotation',
    CategoryProduct = 'categorie-produit',
    BankAccounts = 'compte-bancaire',
    Unites = 'unites',
    Marques = 'marques',
    DocumentParameter = 'parametrage-document',
    PeriodComptable = 'periode-comptable',
    Facture = 'facture',
    Avoir = 'avoir',
    Paiement = 'paiement',
    Comptabilite = 'comptabilite',
    ModeReglement = 'mode-reglement',
    SpecialArticle = 'special-article',
    Devis = 'devis',
    ConfigurationAgenda = 'configuration-agenda',
    TypeLogement = 'type-logement',
    Profile = 'utilisateurs/profile',
    ModeleSMS = 'modele-sms',
    AgendaCommercial = 'agenda-commercial',
    TypeChauffage = 'type-chauffage',
    SiteInstallation = 'site-installation',
    SuiviActivite = 'suivi-activite',
    ChampSiteInstallation = 'champs-site-intervention',
    BonCommande = 'bon-commande',
    SourceDuLead = 'source-du-lead',
}